app/requirements.txt
mysqlclient
mysql-connector-python
boto3